Refactoring file hierarchy
