+++ src/ram/TranslationUnit.h
@@ -17,10 +17,10 @@
 
 #pragma once
 
-#include "DebugReport.h"
 #include "Global.h"
 #include "ram/Program.h"
 #include "ram/analysis/Analysis.h"
+#include "reports/DebugReport.h"
 #include "souffle/SymbolTable.h"
 #include <cassert>
 #include <iosfwd>
