+++ src/tests/ast_print_test.cpp
@@ -17,8 +17,6 @@
 #include "tests/test.h"
 
 #include "AggregateOp.h"
-#include "DebugReport.h"
-#include "ErrorReport.h"
 #include "ast/Argument.h"
 #include "ast/Atom.h"
 #include "ast/Clause.h"
@@ -28,6 +26,8 @@
 #include "ast/TranslationUnit.h"
 #include "ast/Variable.h"
 #include "parser/ParserDriver.h"
+#include "reports/DebugReport.h"
+#include "reports/ErrorReport.h"
 #include <algorithm>
 #include <iosfwd>
 #include <memory>
