+++ src/tests/ram_relation_test.cpp
@@ -16,8 +16,6 @@
 
 #include "tests/test.h"
 
-#include "DebugReport.h"
-#include "ErrorReport.h"
 #include "Global.h"
 #include "RelationTag.h"
 #include "interpreter/InterpreterEngine.h"
@@ -31,6 +29,8 @@
 #include "ram/SignedConstant.h"
 #include "ram/Statement.h"
 #include "ram/TranslationUnit.h"
+#include "reports/DebugReport.h"
+#include "reports/ErrorReport.h"
 #include "souffle/RamTypes.h"
 #include "souffle/SymbolTable.h"
 #include "souffle/json11.h"
